sensor as location migration and removing sensor id from model schema

+++ packages/api/db/migration/20220715165156_sensor_as_standard_location.js
@@ -1,20 +1,5 @@
 exports.up = async function (knex) {
-  // delete sensor table data
-  const sensorLocationIds = await knex('sensor').select('location_id');
-
-  if (sensorLocationIds.length > 0) {
-    await knex.raw(`
-            BEGIN TRANSACTION;
-            DELETE FROM point p WHERE p.figure_id IN
-            (SELECT f.figure_id FROM figure f WHERE f.type = 'sensor');
-            DELETE FROM figure f WHERE f.type = 'sensor';
-            DELETE FROM sensor_reading;
-            DELETE FROM sensor_reading_type;
-            DELETE FROM sensor;
-            COMMIT;
-        `);
-    await knex('location').whereIn('location_id', sensorLocationIds).del();
-  }
+  await deleteSensorData(knex);
 
   await knex.raw(`
         DROP TABLE sensor_reading;
@@ -45,7 +30,7 @@ exports.up = async function (knex) {
   await knex.schema.createTable('sensor_reading', function (table) {
     table.uuid('reading_id').primary().notNullable().defaultTo(knex.raw('uuid_generate_v1()'));
     table
-      .string('location_id')
+      .uuid('location_id')
       .notNullable()
       .references('location_id')
       .inTable('sensor')
@@ -72,18 +57,26 @@ exports.up = async function (knex) {
   });
 };
 
-exports.down = function (knex) {
-  // delete sensor table data
-  knex.raw(`
-  BEGIN TRANSACTION;
-  DELETE FROM point p WHERE p.figure_id IN
-  (SELECT f.figure_id FROM figure f WHERE f.type = 'sensor');
-  DELETE FROM figure f WHERE WHERE f.type = 'sensor';
-  DELETE FROM location l WHERE l.location_id IN
-  (SELECT s.location_id FROM sensor);
-  DROP TABLE sensor_reading;
-  DROP TABLE sensor_reading_type;
-  DROP TABLE sensor;
-  COMMIT;
-`);
+exports.down = async function (knex) {
+  await deleteSensorData(knex);
 };
+
+async function deleteSensorData(knex) {
+  const sensorLocationIdObjs = await knex('sensor').select('location_id');
+  const sensorLocationIds = sensorLocationIdObjs.map((s) => s.location_id);
+
+  if (sensorLocationIds.length > 0) {
+    await knex.raw(`
+              BEGIN TRANSACTION;
+              DELETE FROM point p WHERE p.figure_id IN
+              (SELECT f.figure_id FROM figure f WHERE f.type = 'sensor');
+              DELETE FROM figure f WHERE f.type = 'sensor';
+              DELETE FROM sensor_reading;
+              DELETE FROM sensor_reading_type;
+              DELETE FROM sensor;
+              COMMIT;
+          `);
+
+    await knex('location').whereIn('location_id', sensorLocationIds).del();
+  }
+}
